+++ backend/src/main/java/com/weread/entity/user/FollowId.java
@@ -6,6 +6,6 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 @Data
 public class FollowId { // 放在 FollowId.java 文件中，所以可以是 public
-    private Long followerId;  
-    private Long followedId;  
+    private Integer followerId;  
+    private Integer followedId;  
 }
