+++ backend/src/main/java/com/weread/repository/book/UserBookAccessRepository.java
@@ -4,5 +4,5 @@ import com.weread.entity.book.UserBookAccessEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserBookAccessRepository extends JpaRepository<UserBookAccessEntity, Long> {
-    boolean existsByUserIdAndBookId(Long userId, Long bookId);
+    boolean existsByUserIdAndBookId(Long userId, Integer bookId);
 }
